requirements.txt
absl-py==2.3.0
appdirs==1.4.4
asttokens==2.4.1
backcall==0.2.0
cached-property==2.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.1
crafter==1.8.0
cycler==0.12.1
Cython==0.29.37
debugpy==1.6.7
decorator==4.4.2
dm-control==1.0.9
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.3.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.15
filelock==3.18.0
flatten-dict==0.4.2
fonttools==4.58.4
free-mujoco-py==2.1.6
fsspec==2025.5.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.5.5
grpcio==1.73.1
gym==0.22.0
gym-notices==0.0.8
h5py==3.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.29.3
ipython==8.15.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.5.2
jupyter-client==7.3.4
jupyter_core==5.7.2
kiwisolver==1.4.7
kornia==0.8.1
kornia_rs==0.1.9
labmaze==1.0.6
llvmlite==0.43.0
lxml==6.0.0
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
memory-maze==1.0.3
metaworld==0.0.0
mj_envs==1.0.0
mjrl==1.0.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.3.5
mujoco-py==2.1.2.14
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.7.0.72
opensimplex==0.4.5.1
packaging==24.2
pandas==2.3.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==23.3.1
platformdirs==4.3.7
pluggy==1.6.0
proglog==0.1.12
prompt-toolkit==3.0.42
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pygame==2.6.1
Pygments==2.17.2
PyOpenGL==3.1.9
pyparsing==3.2.3
pyqt5==5.15.9
pyrender==0.1.45
pyrep==4.1.0.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.1.2
requests==2.32.4
RRL==1.0.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.12
scikit-video==1.1.11
scipy==1.13.1
sentry-sdk==2.32.0
setproctitle==1.3.6
setuptools==78.1.1
setuptools-scm==8.3.1
six==1.16.0
smmap==5.0.2
stack-data==0.6.2
sympy==1.14.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==3.1.0
tomli==2.2.1
torch==2.1.0
torchvision==0.16.0
tornado==6.1
tqdm==4.67.1
traitlets==5.14.3
triton==2.1.0
typing_extensions==4.10.0
urllib3==2.5.0
wandb==0.13.10
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
xmltodict==0.13.0
yt-dlp==2025.6.30
zipp==3.23.0
